+++ lib/tests/wed/gui/dialog-search-replace-test.ts
@@ -19,9 +19,7 @@ const expect = chai.expect;
 describe("dialog-search-replace", () => {
   let setup: EditorSetup;
   let editor: Editor;
-  let guiRoot: HTMLElement;
   let dataRoot: Document;
-  let docScope: DLocRange;
   let caretManager: CaretManager;
 
   let ps: Element[];
@@ -33,10 +31,6 @@ describe("dialog-search-replace", () => {
   let firstABCText: DLocRange;
   let firstABCDText: DLocRange;
   let firstABCAttribute: DLocRange;
-  let firstABCDAttribute: DLocRange;
-  let secondABCAttribute: DLocRange;
-  let titleBCD: DLocRange;
-  let titleABCD: DLocRange;
 
   before(() => {
     setup = new EditorSetup(
@@ -48,18 +42,8 @@ describe("dialog-search-replace", () => {
   });
 
   beforeEach(() => {
-    guiRoot = editor.guiRoot;
     dataRoot = editor.dataRoot;
     caretManager = editor.caretManager;
-    docScope = editor.caretManager.docDLocRange;
-
-    const title = dataRoot.querySelector("title")!;
-    titleBCD = new DLocRange(
-      caretManager.mustFromDataLocation(title.firstChild!, 1),
-      caretManager.mustFromDataLocation(title.firstChild!, 4));
-    titleABCD = new DLocRange(
-      caretManager.mustFromDataLocation(title.firstChild!, 0),
-      caretManager.mustFromDataLocation(title.firstChild!, 4));
 
     ps = Array.from(dataRoot.querySelectorAll("body p"));
     firstBodyP = ps[0];
@@ -95,18 +79,10 @@ describe("dialog-search-replace", () => {
       caretManager.mustFromDataLocation(ps[3].firstChild!.firstChild!, 0),
       caretManager.mustFromDataLocation(ps[3].lastChild!, 2));
 
-    const rend = ps[7].getAttributeNode("rend");
+    const rend = ps[7].getAttributeNode("rend")!;
     firstABCAttribute = new DLocRange(
       caretManager.mustFromDataLocation(rend, 0),
       caretManager.mustFromDataLocation(rend, 3));
-
-    firstABCDAttribute = new DLocRange(
-      caretManager.mustFromDataLocation(rend, 0),
-      caretManager.mustFromDataLocation(rend, 4));
-
-    secondABCAttribute = new DLocRange(
-      caretManager.mustFromDataLocation(rend, 4),
-      caretManager.mustFromDataLocation(rend, 7));
   });
 
   afterEach(() => {